+++ banner/src/main/res/drawable/mount_fuji_afternoon.xml
