+++ banner/src/main/res/drawable/mount_fuji_night.xml
