+++ settings/build.gradle.kts
@@ -36,6 +36,7 @@ dependencies {
     implementation(projects.api.anilist)
     implementation(projects.api.preferences)
     implementation(projects.core)
+    implementation(projects.banner)
     implementation(projects.media)
     implementation(projects.navigation)
